requirements.txt
dotmap==1.3.26
imgaug==0.4.0
matplotlib==3.3.4
pandas==1.1.5
pydot==1.4.2
PyYAML==5.4.1
scikit-learn==0.24.2
seaborn==0.11.2
tensorflow==2.7.0